RTech: fix numeric overflow caused by massive pak files
Pak files above 4GB caused the async file system to throw an error about failing to load audio banks, removed casts and promoted everything correctly. Also removed irrelevant comments in the public pakfile header.

+++ src/rtech/async/asyncio.cpp
@@ -55,7 +55,7 @@ int FS_OpenAsyncFile(const char* const filePath, const int logLevel, size_t* con
 //----------------------------------------------------------------------------------
 // close a file and remove it from the async file handle array
 //----------------------------------------------------------------------------------
-void FS_CloseAsyncFile(const short fileHandle)
+void FS_CloseAsyncFile(const int fileHandle)
 {
     const int slotNum = fileHandle & ASYNC_MAX_FILE_HANDLES_MASK;
     AsyncHandleTracker_t& tracker = g_pAsyncFileSlots[slotNum];
